% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmam.R
\name{bmam}
\alias{bmam}
\title{Fit Bayesian Marginal Additive Model (BMAM)}
\usage{
bmam(
  object,
  preddat,
  length = 100,
  summarize = TRUE,
  posterior = TRUE,
  backtrans = c("response", "linear", "identity", "invlogit", "exp", "square", "inverse"),
  centered = FALSE,
  k = 100,
  ...
)
}
\arguments{
\item{object}{A fitted brms model object that includes random effects.
Required.}

\item{preddat}{Predicted data. A data frame containing covariates at which
predictions are required. If empty, it will be generated by
\code{generate_pred()}}

\item{length}{Number of observations in the generated predicted data.}

\item{summarize}{A logical value, whether or not to calculate summaries of
the posterior predictions. Defaults to \code{TRUE}.}

\item{posterior}{A logical value whether or not to save and return the
posterior samples. Defaults to \code{FALSE} as the assumption is a typical
use case is to return the summaries only.}

\item{backtrans}{A character string indicating the type of back
transformation to be applied. Can be one of \dQuote{response} meaning to
use the response scale, \dQuote{linear} or \dQuote{identity} meaning to use
the linear predictor scale, or a specific back transformation desired, from
a possible list of \dQuote{invlogit}, \dQuote{exp}, \dQuote{square}, or
\dQuote{inverse}. Custom back transformations should only be needed if, for
example, the outcome variable was transformed prior to fitting the model.
default: \dQuote{response}}

\item{centered}{Whether or not return centered smooths. Default \code{FALSE}
If \code{TRUE}, the smooth predictions returned will have their means
subtracted such that they have mean zero across the supplied covariate
values for predictions. See \code{MAM} package.}

\item{k}{An integer providing the number of random draws to use for
integrating out the random effects. Only relevant when \code{effects} is
\dQuote{integrateoutRE}.}

\item{...}{Additional arguments passed to \code{fitted()}}
}
\value{
Objects of Class 'bmam', a list containing the fitted Bayesian
  marginal additive model and the results from the conditional model. Some of
  these may be \code{NULL} depending on the arguments used.
}
\description{
Fit Bayesian Marginal Additive Model (BMAM). The function is the modification
of \code{brmsmargins::marginalcoef}.
}
\references{
Hedeker, D., du Toit, S. H., Demirtas, H. & Gibbons, R. D. (2018)
  \doi{10.1111/biom.12707} \dQuote{A note on marginalization of regression
  parameters from mixed models of binary outcomes}

  McGee G. & Stringer A. (2018) \dQuote{Flexible Marginal Models for
  Dependent Data}
}
\seealso{
\code{\link{brms}}, \code{\link{brmsmargins}}, \code{\link{mam}}
}
