% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{bsummary}
\alias{bsummary}
\title{Bayesian Summary. Function in brmsmargins. Change default.
See ?brmsmargins::bsummary}
\usage{
bsummary(x, CI = 0.95, CIType = "ETI", ...)
}
\arguments{
\item{x}{The posterior samples of a parameter}

\item{CI}{Width of the credible interval.}

\item{CIType}{Type of credible interval, passed on
to the \code{\link[bayestestR]{ci}} function as the method for CIs.}
}
\description{
Bayesian Summary. Function in brmsmargins. Change default.
See ?brmsmargins::bsummary
}
\seealso{
\code{\link{brmsmargins::bsummary}}
}
