% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders_hs.R
\name{builders}
\alias{builders}
\alias{.namesL_hs}
\alias{.buildL_hs}
\alias{.namesSD_hs}
\alias{.buildSD_hs}
\alias{.namesZ_hs}
\alias{.buildZ_hs}
\title{Build the Variable Names or Data Objects for Estimation
Modification from builder.r in brmsmargins
For horseshoe prior}
\usage{
.namesL_hs(ranef, nlpar)

.buildL_hs(data, ranef, nlpar)

.namesSD_hs(ranef, block)

.buildSD_hs(data, ranef, block)

.namesZ_hs(block, number, nlpar)

.buildZ_hs(data, block, number, nlpar)
}
\arguments{
\item{ranef}{A data set with information about the model object random effects.
Only used for \code{.namesSD} and \code{.buildSD}.}

\item{data}{A data object. For example the result of [make_standata()]
for [.buildZ()], which is a list,
or a dataset of the posterior draws such as from [as_draws_df()]
for [.buildL()] and [.buildSD()].}

\item{block}{Which random effect block to use. An integer.}

\item{number}{The number of elements in that random effect block. An integer.}

\item{dpar}{Which dpar to use. Does not apply to the L matrix.}
}
\value{
A character vector for all \code{.names} functions or a matrix
  for all \code{.build} functions.
}
\description{
These are a set of internal utility functions.
They are not intended for general use.
}
\details{
\itemize{
  \item{\code{.namesL_hs}}{Generate names of an L matrix from \code{brms}. Create the variable names for the Cholesky decomposition of the random effects correlation matrix in \code{brms}. Note that \code{brms} returns the lower triangular matrix and we want the upper triangular matrix, so the names are transposed. The results can then be passed to the \code{tab2mat} function to convert the row vector into a matrix.}
  \item{\code{.buildL_hs}}{Returns the L matrix object. Rows are posterior draws.}
  \item{\code{.namesSD_hs}}{Create the names of random effect standard deviation estimates.}
  \item{\code{.buildSD_hs}}{Return matrix of random effect standard deviation estimates. Rows are posterior draws.}
  \item{\code{.namesZ_hs}}{Create the names of random effects data for predictions.}
  \item{\code{.buildZ_hs}}{Return matrix of data for random effect predictions.}
}
}
\keyword{internal}
