% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_brms.R
\name{conditional_brms}
\alias{conditional_brms}
\title{Prediction from conditional GAMM}
\usage{
conditional_brms(object, data, centered = FALSE, ...)
}
\arguments{
\item{object}{object Objects of Class 'brms'}

\item{data}{A data frame containing covariates at which predictions are required.}

\item{centered}{whether or not return centered smooths. Default \code{FALSE}
If \code{TRUE}, the smooth predictions returned will have their means
subtracted such that they have mean zero across the supplied covariate
values for predictions. See \code{MAM} package.}

\item{...}{Additional arguments passed to \code{bsummary()}}
}
\value{
a data frame containing prediction results (\code{bsummary()}) from the conditional GAMM
}
\description{
Prediction from conditional GAMM
}
