% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenBinaryY.R
\name{Etai2Deltai1}
\alias{Etai2Deltai1}
\title{Generate binary response variable of GLMM model based on a known Marginalized model}
\usage{
Etai2Deltai1(etai, gamma, sigma, q.points, Z, W)
}
\description{
Generate binary response variable of GLMM model based on a known Marginalized model
}
