% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenBinaryY.R
\name{GenBinaryY}
\alias{GenBinaryY}
\title{Generate binary Y}
\usage{
GenBinaryY(
  mean.formula,
  lv.formula = NULL,
  t.formula = NULL,
  beta = NULL,
  Sigma = NULL,
  gamma = NULL,
  id,
  data,
  q = 10,
  Yname = "Y"
)
}
\arguments{
\item{lv.formula}{lv formula for GLMM model}

\item{t.formula}{transform (lag term) for GLMM model}

\item{beta}{coefficients for marginalized model}

\item{Sigma}{covariance matrix for random effects}

\item{gamma}{coefficients for lag term}

\item{id}{id for clusters}

\item{data}{a dataframe containing explanatory variables and id}

\item{q}{number of nodes}

\item{Yname}{name of generated reponse variable in the returned dataframe}
}
\value{
a dataframe contains explanatory variables, id and generated binary
  response variable
}
\description{
Generate binary Y
}
\note{
the function reduce the \eqn{X\gamma} into one-dim normal distributed
  variable, and then using Gauss-Hermite Quadrature to solve delta. After
  getting the solution, generate data based on \eqn{X\gamma} but not the
  one-dim term. This function need include the R file Etai2Deltai1.R,
  get.GH.R, expit.R and C code source from src1.so. See more details in
  GenBinY_Summary.md and 1 dim in LME.md
}
\author{
Tianyi Pan changing the GenBinaryY function in MMLB
  package(https://github.com/mercaldo/MMLB).
}
