% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pred.R
\name{generate_pred}
\alias{generate_pred}
\title{generate predicted data}
\usage{
generate_pred(object, length = 100, hsformula)
}
\arguments{
\item{object}{Objects of Class 'brms'}

\item{length}{number of observations in the generated data}
}
\value{
predicted data, used by \code{marginalcoef()}
}
\description{
generate predicted data
}
