% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bmam.R
\name{plot.bmamfit}
\alias{plot.bmamfit}
\title{plot() Method for Objects of Class 'bmamfit'}
\usage{
\method{plot}{bmamfit}(
  object,
  compared.model,
  conditional = TRUE,
  display = TRUE,
  smooth.function
)
}
\arguments{
\item{object}{Objects of Class 'bmamfit'}

\item{compared.model}{Other model compared with BMAM.
supported models: 1. mam
                  2. gam
                  3. brms gam}

\item{conditional}{Whether or not to plot the conditional model. Default: TRUE}

\item{display}{Whether or not to display the plot. Default: TRUE}

\item{smooth.function}{A list. True values of the smooth functions.}
}
\value{
a list containing ggplot objects.
}
\description{
plot() Method for Objects of Class 'bmamfit'
}
