% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.bmamfit}
\alias{print.bmamfit}
\title{print() Method for Objects of Class 'bmamfit'}
\usage{
\method{print}{bmamfit}(object, ...)
}
\arguments{
\item{object}{Objects of Class 'bmamfit'}
}
\value{
invisible() with printing
}
\description{
print() Method for Objects of Class 'bmamfit'
}
