% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimData}
\alias{SimData}
\title{Generate simulated data}
\usage{
SimData(K, Nk)
}
\arguments{
\item{K}{Number of clusters}

\item{Nk}{Number of units within clusters}
}
\value{
a list containing generated data and true functions.
}
\description{
Generate simulated data
}
